requirements.txt
streamlit==1.31.0
ultralytics==8.1.19
opencv-python==4.9.0.80
Pillow==10.2.0
scikit-learn==1.4.0
torch==2.2.0
numpy==1.26.3
pandas==2.2.0